Rename TabbedBrowser.shutting_down
It was made public in 1c10a1aecfd50 and the method that used it
externally was removed in 08bb3f4f19b51.

I renamed it to be the same as in the views and so I can add a
shutting_down signal.

As an aside the `Application.new_window` signal that was also added for
1c10a1aecfd50 also had its sole consumer removed in that refactoring. I
can see more future use of it though.