Maybe fix tests.
Getting a failure on travis where the window created by undo -w was not
active. Hopefully it just needs to wait more and it isn't semothing like
the session being save earlier.

Also sometimes after closing a window the next command will be send
after the window has been destroyed but before it was removed from the
window registry. So the IPC server tries to run the command in a deleted
window, watch for that.